Added 'karl.scripting.only_once' API, to avoid overrunning cron-driven scripts.
The overrun detection is based around the existence of a lockfile named
'<username>-<progname>', where '<username> is the current user's login name,
and '<progname>' is a string passed to 'only_once'. 

By default, the lockfile is created in '/var/run/karl/'; this path can be
overridden by setting 'lockdir' in the 'DEFAULT' section of the configuration
file.