Make sure that tinymce does not try to load any plugin css
Tinymce's current way of loading its plugin css breaks
respond.js which provides responsive designs to work
on browsers that lack support for it, such as IE<=8.

While there is no fix known to this problem yet,
we can avoid it in KARL because we load the tinymce
css statically. Before the current fix
this was broken which exposed us to
the original problem.