Hold the jQuery ready event to be triggered, until all our libraries are loaded
Since we are deferring the loading of our libraries,
it may happen that the jQuery document ready event is
fired before some or all of the library code is
available. This only happens on some browser and sometimes.
To make sure the ready event does not trigger earlier
then we want, we hold the event until the last of
our deferred script has been loaded.

As a remark, this may be the main reason that accounts
for previously observed IE8 anomalies.