Make sure selectivizr (on IE8) is loaded before any css.
A nasty bug occurs with selectivizr under the following combination:

- a css is loaded before the selectivizr.js
- in this css there is an @import xxx.css with no trailing end of line
- if this was generated from .less, the error only occurs if the
 file is compressed _and_ loads a real css directly.

A consequence is that selectivizr will try to treat the
entire css content as an url, resulting an Access Denied in line 239.

The situation is prevented on IE8 if we make sure the selectivizr.js
is loaded prior to any css files.

Fixes LP #1009137.