Fix the community id index
Indxes involving functions that do queries is a no-no and can cause
index corruption, as has happened with the community id index.

This set of evolutions creates a trigger that stores the community id
as records are inserted or updated.  It then replaces the current
get_community_zoid function and index to use this stored data. This
has been tested in stage.